+++ extension-compose/src/main/java/com/mapbox/maps/extension/compose/style/layers/internal/LayerNode.kt
@@ -2,12 +2,12 @@ package com.mapbox.maps.extension.compose.style.layers.internal
 
 import androidx.compose.runtime.Stable
 import com.mapbox.bindgen.Value
+import com.mapbox.maps.MapboxExperimental
 import com.mapbox.maps.MapboxMap
-import com.mapbox.maps.MapboxStyleManager
 import com.mapbox.maps.extension.compose.internal.LayerPositionAwareNode
 import com.mapbox.maps.extension.compose.internal.MapNode
+import com.mapbox.maps.extension.compose.internal.StyleLifecycleAwareNode
 import com.mapbox.maps.extension.compose.style.StyleImage
-import com.mapbox.maps.extension.compose.style.internal.StyleAwareNode
 import com.mapbox.maps.extension.compose.style.internal.StyleLayerPositionNode
 import com.mapbox.maps.extension.compose.style.internal.StyleSlotNode
 import com.mapbox.maps.extension.compose.style.sources.SourceState
@@ -15,11 +15,6 @@ import com.mapbox.maps.logD
 import com.mapbox.maps.logE
 import com.mapbox.maps.logW
 import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.filterNotNull
-import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.flow.firstOrNull
-import kotlinx.coroutines.launch
 
 @Stable
 internal class LayerNode(
@@ -27,8 +22,8 @@ internal class LayerNode(
   private val layerType: String,
   private var layerId: String,
   private var sourceState: SourceState? = null,
-  private val coroutineScope: CoroutineScope,
-) : LayerPositionAwareNode, MapNode() {
+  coroutineScope: CoroutineScope,
+) : LayerPositionAwareNode, StyleLifecycleAwareNode(map, coroutineScope) {
   private val parameters = hashMapOf(
     "id" to Value(layerId),
     "type" to Value(layerType)
@@ -38,48 +33,26 @@ internal class LayerNode(
     }
   }
 
-  private val styleManagerFlow: MutableStateFlow<MapboxStyleManager?> = MutableStateFlow(null)
-
   private val addedImages = mutableListOf<String>()
   private val addedModels = mutableListOf<String>()
 
   private lateinit var parentNode: MapNode
 
+  override var isAttached: Boolean = false
+
   override fun onAttached(parent: MapNode) {
     logD(TAG, "onAttached: parent=$parent")
-    parameters.apply {
-      if (parent is StyleSlotNode) {
-        this["slot"] = Value(parent.slotName)
-      }
+    super.onAttached(parent)
+    if (parent is StyleSlotNode) {
+      parameters["slot"] = Value(parent.slotName)
     }
     parentNode = parent
-    updateStyleManagerState(parent)
     addLayer()
   }
 
-  private fun updateStyleManagerState(parent: MapNode) {
-    when (parent) {
-      is StyleAwareNode -> {
-        coroutineScope.launch {
-          parent.mapStyleNode.styleDataLoaded.firstOrNull()?.let {
-            styleManagerFlow.value = map
-          }
-        }
-      }
-
-      else -> styleManagerFlow.value = map
-    }
-  }
-
-  private fun dispatchWhenStyleDataLoaded(action: (MapboxStyleManager) -> Unit) {
-    coroutineScope.launch {
-      styleManagerFlow.filterNotNull().first().apply(action)
-    }
-  }
-
-  override fun onMoved(parent: MapNode, from: Int, to: Int) {
-    logD(TAG, "onMoved: from $from to $to")
-    dispatchWhenStyleDataLoaded {
+  override fun onMoved(parent: MapNode) {
+    logD(TAG, "onMoved")
+    dispatchWhenReady {
       it.repositionCurrentNode(parent)
     }
   }
@@ -110,6 +83,7 @@ internal class LayerNode(
     cleanUpResources()
   }
 
+  @OptIn(MapboxExperimental::class)
   private fun cleanUpResources() {
     addedImages.forEach {
       map.removeStyleImage(it)
@@ -121,64 +95,39 @@ internal class LayerNode(
     addedModels.clear()
   }
 
-  private fun addLayer(parent: MapNode = parentNode) {
-    when (parent) {
-      is StyleSlotNode -> {
-        // layer added within the style node, add as non-persistent layer
-        logD(TAG, "Adding layer: $parameters, at slot: ${parent.slotName}")
-        // fetch the relative position info synchronously as the add layer can be done
-        // async, and the relative position can change later.
-        val layerPosition = getRelativePositionInfo(parent)
-        dispatchWhenStyleDataLoaded {
-          it.addStyleLayer(
-            parameters = Value(parameters),
-            position = layerPosition
-          ).onError {
-            logE(TAG, "Failed to add layer: $it")
-          }.onValue {
-            logD(TAG, "Added layer: $parameters")
-            attachSource()
-          }
-        }
-      }
-
-      is StyleLayerPositionNode -> {
-        // layer added within the style node, add as non-persistent layer
-        logD(TAG, "Adding layer: $parameters, at position: ${parent.layerPosition}")
-        // Add layer to the position defined in [StyleLayerPositionNode] if it's the last [LayerPositionAwareNode]
-        // under the [StyleLayerPositionNode], otherwise, we insert with relative position.
-        val layerPosition = getRelativePositionInfo(parent, parent.layerPosition)
-        dispatchWhenStyleDataLoaded { styleManager ->
-          styleManager.addStyleLayer(
-            parameters = Value(parameters),
-            position = layerPosition
-          ).onError {
-            logE(TAG, "Failed to add layer $parameters at $layerPosition: $it")
-            logE(TAG, "Available layers in style:")
-            styleManager.styleLayers.forEach { logE(TAG, "\t ${it.id}") }
-          }.onValue {
-            logD(TAG, "Added layer: $parameters")
-            attachSource()
-          }
-        }
-      }
-
-      else -> {
-        // layer added outside of the style node, add as persistent layer
-        logD(TAG, "Adding persistent layer: $parameters")
-        // fetch the relative position info synchronously as the add layer can be done
-        // async, and the relative position can change later.
-        val layerPosition = getRelativePositionInfo(parent)
-        map.addPersistentStyleLayer(
-          properties = Value(parameters),
-          layerPosition = layerPosition
-        ).onError {
-          logE(TAG, "Failed to add persistent layer: $it")
+  private fun addLayer() {
+    logD(TAG, "Adding layer: $parameters, within parent: $parentNode")
+    // layer added within the style or slot node, add as non-persistent layer
+    if (parentNode is StyleSlotNode || parentNode is StyleLayerPositionNode) {
+      dispatchWhenReady {
+        val layerPosition = getRelativePositionInfo(parentNode)
+        logD(TAG, "Adding layer: $parameters, with layer position: $layerPosition")
+        it.addStyleLayer(
+          parameters = Value(parameters),
+          position = layerPosition
+        ).onError { error ->
+          logE(TAG, "Failed to add layer: $error")
         }.onValue {
-          logD(TAG, "Added persistent layer: $parameters")
+          logD(TAG, "Added layer: $parameters")
           attachSource()
         }
+        isAttached = true
       }
+    } else {
+      // layer added outside of the style node, add as persistent layer
+      logD(TAG, "Adding persistent layer: $parameters")
+      val layerPosition = getRelativePositionInfo(parentNode)
+      logD(TAG, "Adding layer: $parameters, with layer position: $layerPosition")
+      map.addPersistentStyleLayer(
+        properties = Value(parameters),
+        layerPosition = layerPosition
+      ).onError {
+        logE(TAG, "Failed to add persistent layer: $it")
+      }.onValue {
+        logD(TAG, "Added persistent layer: $parameters")
+        attachSource()
+      }
+      isAttached = true
     }
   }
 
@@ -210,11 +159,12 @@ internal class LayerNode(
     attachSource()
   }
 
+  @OptIn(MapboxExperimental::class)
   internal fun addModel(modelInfo: Pair<String, String>) {
-    dispatchWhenStyleDataLoaded {
+    dispatchWhenReady {
       it.addStyleModel(modelInfo.first, modelInfo.second)
-        .onError {
-          logW(TAG, "Failed to add style model $modelInfo: $it")
+        .onError { error ->
+          logW(TAG, "Failed to add style model $modelInfo: $error")
         }.onValue {
           addedModels.add(modelInfo.first)
         }
@@ -222,7 +172,7 @@ internal class LayerNode(
   }
 
   internal fun addImage(styleImage: StyleImage) {
-    dispatchWhenStyleDataLoaded {
+    dispatchWhenReady {
       it.addStyleImage(
         imageId = styleImage.imageId,
         scale = styleImage.scale ?: it.pixelRatio,
@@ -231,8 +181,8 @@ internal class LayerNode(
         stretchX = styleImage.stretchX,
         stretchY = styleImage.stretchY,
         content = styleImage.content
-      ).onError {
-        logW(TAG, "Failed to add style image $styleImage: $it")
+      ).onError { error ->
+        logW(TAG, "Failed to add style image $styleImage: $error")
       }.onValue {
         addedImages.add(styleImage.imageId)
       }
@@ -241,12 +191,17 @@ internal class LayerNode(
 
   internal fun setProperty(name: String, value: Value) {
     logD(TAG, "[$layerType] settingProperty: property=$name, value=$value ...")
-    dispatchWhenStyleDataLoaded {
+    if (!isAttached) {
+      logD(TAG, "setProperty() called before the layer is attached to the map.")
       parameters[name] = value
-      it.setStyleLayerProperty(layerId, name, value).error?.let {
+      return
+    }
+    dispatchWhenReady {
+      parameters[name] = value
+      it.setStyleLayerProperty(layerId, name, value).error?.let { error ->
         logW(
           TAG,
-          "Failed to set $name property as $value on $layerType layer $layerId: $it"
+          "Failed to set $name property as $value on $layerType layer $layerId: $error"
         )
       }
       logD(TAG, "[$layerType] setProperty: property=$name, value=$value executed")